fix: validate gamestate with agents
Summary:
I would like to use the multi agent fuzzing to test the more precise
action log. So this is a first pass at adding that and cleaning up
around it

Then when writing that fuzz we can re-use the test utils everywhere.

That found an issue. We were transitioning the game to complete if an
agent folded. However there was still one player to act in a previous
round. So since we reset counts and data, it allowed invalid actions

This also found another issue: We were not starting on the dealer flop
on.

This also found another issue: The action log was using an incorrect idx
for the player.

This also found another issue money in the pot can be left orphaned if
the last players all fold.

Test Plan:
The whole thing is a test