feat: add on `CardBitSet` for card sets
Summary:
We only have 52 cards. So we can use the 64 bit integers as flag set.
This does that and makes deck follow the same api. As such this is a
breaking change to the deck struct.

It's recomended to use  CardBitSet whenever possible as it will be
faster and smaller.

Test Plan:
Lots of tests added