fix: allow simulations to restart from player actions
Summary:
When given a game state a simulation should always restart from the
exact same place. This fixes the issues where we were not living up to
that

- We would double blind
- We would not respect that the last player to act in a round might be
  because we are restarting the simulation

This also splits the simulation up into a lot more rounds to make it
easier to see round changes as state changes in the game state.

Test Plan:

This was all found by the added test
Then the solution was further refined with fuzz testing