feat: create StateStore
Summary:
Each `Agent` uses a CFRState to find the optimal solution. This PR
creates a `StateStore` to be a globally accessably (via the player
index). This allows each player to re-use the exploration time and not
allocate new `CFRStates`.

In other words. Previously when each player would be asked to choose the
best action. We would fire off a lot of games to simulate and the agents
we were playing against would need to have their own `CFRStates`. This
caused a lot of memory churn.

Since `CFRState` is an pointer any way it's much less merory to recreate
on each exploration attempt

Additional Changes:

- Moved to svg for default in example
- Tests Added